int to size_t

+++ XOR/matrix.h
@@ -3,5 +3,5 @@
 double initWeights();
 double sigmoid(double x);
 double sigmoidDerivative(double x);
-void shuffle(int *array, int n);
+void shuffle(int *array, size_t n);
 int main();
